Move chimp config into a file.
