+++ .travis.yml
@@ -5,11 +5,9 @@ node_js:
 cache: yarn
 
 before_install:
-  - sudo apt-get -qq update
-  - sudo apt-get install -y oracle-java8-set-default
-  - export CHROME_BIN=chromium-browser
   - export DISPLAY=:99.0
   - sh -e /etc/init.d/xvfb start
+  - sleep 3 # give xvfb some time to start
 
 install:
  - yarn install
